remove cdc-query-related codes
