+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/hudi/HoodieDataSourceHelper.scala
@@ -18,17 +18,13 @@
 
 package org.apache.hudi
 
-import org.apache.avro.Schema
-import org.apache.avro.generic.GenericRecord
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.FileStatus
 import org.apache.hudi.client.utils.SparkInternalSchemaConverter
 import org.apache.hudi.common.util.StringUtils.isNullOrEmpty
-import org.apache.hudi.common.util.ValidationUtils.checkState
 import org.apache.hudi.internal.schema.InternalSchema
 import org.apache.hudi.internal.schema.utils.SerDeHelper
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.avro.HoodieAvroDeserializer
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{PredicateHelper, SpecificInternalRow, UnsafeProjection}
 import org.apache.spark.sql.execution.datasources.PartitionedFile
@@ -86,19 +82,6 @@ object HoodieDataSourceHelper extends PredicateHelper with SparkAdapterSupport {
       val remaining = file.getLen - offset
       val size = if (remaining > maxSplitBytes) maxSplitBytes else remaining
       PartitionedFile(partitionValues, filePath.toUri.toString, offset, size)
-    }
-  }
-
-  trait AvroDeserializerSupport extends SparkAdapterSupport {
-    protected val avroSchema: Schema
-    protected val structTypeSchema: StructType
-
-    private lazy val deserializer: HoodieAvroDeserializer =
-      sparkAdapter.createAvroDeserializer(avroSchema, structTypeSchema)
-
-    protected def deserialize(avroRecord: GenericRecord): InternalRow = {
-      checkState(avroRecord.getSchema.getFields.size() == structTypeSchema.fields.length)
-      deserializer.deserialize(avroRecord).get.asInstanceOf[InternalRow]
     }
   }
 }
